Error mio solucionado que pena con ustedes
