--- src/views/HTML/servicio_control.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../static/CSS/servicios.css">
-    <title>Servicio de Control</title>
-</head>
-<body>
-    <header class="container">
-        <nav>
-            <a href="/" class="nav-logo">
-                <img src="../static/images/milen-2.png" alt="Logo Milenzo">
-            </a>
-            <div class="nav-buttons">
-                <a href="/myPets" class="nav-button">Mis Mascotas</a>
-                <a href="/serv" class="nav-button">Servicios</a>
-                <a href="/about" class="nav-button">Nosotros</a>
-            </div>
-        </nav>
-    </header>
-
-    <main>
-        <h1>Servicio de Control</h1>
-        <section class="form-section">
-            <h2>Agendar Control</h2>
-            <form id="control-form">
-                <label for="control-mascota">Seleccionar mascota</label>
-                <select id="control-mascota" name="id_pet" required>
-                    <option value="">-- Selecciona una mascota --</option>
-                </select>
-
-                <label for="control-fecha">Fecha</label>
-                <input type="date" id="control-fecha" name="fecha" required>
-
-                <label for="control-hora">Hora</label>
-                <input type="time" id="control-hora" name="hora" required>
-
-                <label for="control">Comentarios</label>
-                <textarea id="control-comentarios" name="comentarios" rows="3" placeholder="Comentarios adicionales..."></textarea>
-
-                <button type="button" id="agendar-control">Agendar Control</button>
-            </form>
-        </section>
-
-        <section class="table-section">
-            <h2>Controles Agendados</h2>
-            <table id="tabla-controles">
-                <thead>
-                    <tr>
-                        <th>Mascota</th>
-                        <th>Fecha</th>
-                        <th>Hora</th>
-                        <th>Comentarios</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <!-- Los baños se llenarán dinámicamente -->
-                </tbody>
-            </table>
-        </section>
-    </main>
-
-    <footer class="footer">
-        <p>&copy; 2025 Milenzo. Todos los derechos reservados.</p>
-    </footer>
-
-    <script src="../static/JS/servicio_control.js"></script>
-</body>
-</html>
