--- src/views/HTML/servicio_corte.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../static/CSS/servicios.css">
-    <title>Servicio de Corte</title>
-</head>
-<body>
-    <header class="container">
-        <nav>
-            <a href="/" class="nav-logo">
-                <img src="../static/images/milen-2.png" alt="Logo Milenzo">
-            </a>
-            <div class="nav-buttons">
-                <a href="/myPets" class="nav-button">Mis Mascotas</a>
-                <a href="/serv" class="nav-button">Servicios</a>
-                <a href="/about" class="nav-button">Nosotros</a>
-            </div>
-        </nav>
-    </header>
-
-    <main>
-        <h1>Servicio de Corte</h1>
-        <section class="form-section">
-            <h2>Agendar Corte</h2>
-            <form id="corte-form">
-                <label for="corte-mascota">Seleccionar mascota</label>
-                <select id="corte-mascota" name="id_pet" required>
-                    <option value="">-- Selecciona una mascota --</option>
-                </select>
-
-                <label for="corte-fecha">Fecha</label>
-                <input type="date" id="corte-fecha" name="fecha" required>
-
-                <label for="corte-hora">Hora</label>
-                <input type="time" id="corte-hora" name="hora" required>
-
-                <label for="corte-comentarios">Comentarios</label>
-                <textarea id="corte-comentarios" name="comentarios" rows="3" placeholder="Comentarios adicionales..."></textarea>
-
-                <button type="button" id="agendar-corte">Agendar Corte</button>
-            </form>
-        </section>
-
-        <section class="table-section">
-            <h2>Cortes Agendados</h2>
-            <table id="tabla-cortes">
-                <thead>
-                    <tr>
-                        <th>Mascota</th>
-                        <th>Fecha</th>
-                        <th>Hora</th>
-                        <th>Comentarios</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <!-- Los baños se llenarán dinámicamente -->
-                </tbody>
-            </table>
-        </section>
-    </main>
-
-    <footer class="footer">
-        <p>&copy; 2025 Milenzo. Todos los derechos reservados.</p>
-    </footer>
-
-    <script src="../static/JS/servicio_corte.js"></script>
-</body>
-</html>
