--- src/views/HTML/servicio_historial.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../static/CSS/servicios.css">
-    <title>Servicio de Historial</title>
-</head>
-<body>
-    <header class="container">
-        <nav>
-            <a href="/" class="nav-logo">
-                <img src="../static/images/milen-2.png" alt="Logo Milenzo">
-            </a>
-            <div class="nav-buttons">
-                <a href="/myPets" class="nav-button">Mis Mascotas</a>
-                <a href="/serv" class="nav-button">Servicios</a>
-                <a href="/about" class="nav-button">Nosotros</a>
-            </div>
-        </nav>
-    </header>
-
-    <main>
-        <h1>Historial Médico</h1>
-        <section class="form-section">
-            <h2>Ver Historial</h2>
-            <form id="Historial-form">
-                <label for="Historial-mascota">Seleccionar mascota</label>
-                <select id="historial-mascota" name="id_pet" required>
-                    <option value="">-- Selecciona una mascota --</option>
-                </select>
-
-                <button type="button" id="ver-historial">Ver Historial</button>
-            </form>
-        </section>
-
-        <section class="table-section">
-            <h2>Historial</h2>
-            <table id="tabla-historial">
-                <thead>
-                    <tr>
-                        <th>Mascota</th>
-                        <th>Fecha</th>
-                        <th>Hora</th>
-                        <th>Comentarios</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <!-- Los baños se llenarán dinámicamente -->
-                </tbody>
-            </table>
-        </section>
-    </main>
-
-    <footer class="footer">
-        <p>&copy; 2025 Milenzo. Todos los derechos reservados.</p>
-    </footer>
-
-    <script src="../static/JS/servicio_historial.js"></script>
-</body>
-</html>
