--- src/views/HTML/servicio_veterinarios.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="../static/CSS/servicios.css">
-    <title>Servicio de Veterinarios</title>
-</head>
-<body>
-    <header class="container">
-        <nav>
-            <a href="/" class="nav-logo">
-                <img src="../static/images/milen-2.png" alt="Logo Milenzo">
-            </a>
-            <div class="nav-buttons">
-                <a href="/myPets" class="nav-button">Mis Mascotas</a>
-                <a href="/serv" class="nav-button">Servicios</a>
-                <a href="/about" class="nav-button">Nosotros</a>
-            </div>
-        </nav>
-    </header>
-
-    <main>
-        <h1>Veterinarios</h1>
-        <section class="form-section">
-            <h2>Contactar veterinarios</h2>
-            <form id="veterinarios-form">
-                <label for="bano-mascota">Seleccionar mascota</label>
-                <select id="bano-mascota" name="id_pet" required>
-                    <option value="">-- Selecciona una mascota --</option>
-                </select>
-
-                <label for="bano-fecha">Fecha</label>
-                <input type="date" id="bano-fecha" name="fecha" required>
-
-                <label for="bano-hora">Hora</label>
-                <input type="time" id="bano-hora" name="hora" required>
-
-                <label for="bano-comentarios">Comentarios</label>
-                <textarea id="bano-comentarios" name="comentarios" rows="3" placeholder="Comentarios adicionales..."></textarea>
-
-                <button type="button" id="agendar-bano">Ver Veterinarios</button>
-            </form>
-        </section>
-
-        <section class="table-section">
-            <h2>Baños Agendados</h2>
-            <table id="tabla-banos">
-                <thead>
-                    <tr>
-                        <th>Mascota</th>
-                        <th>Fecha</th>
-                        <th>Hora</th>
-                        <th>Comentarios</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <!-- Los baños se llenarán dinámicamente -->
-                </tbody>
-            </table>
-        </section>
-    </main>
-
-    <footer class="footer">
-        <p>&copy; 2025 Milenzo. Todos los derechos reservados.</p>
-    </footer>
-
-    <script src="../static/JS/servicio_veterinarios.js"></script>
-</body>
-</html>
